+++ methods/src/test/kotlin/methods/equations/EquationsByFactoringTest.kt
@@ -23,7 +23,7 @@ class EquationsByFactoringTest {
 
         check {
             fromExpr = "[x ^ 2] + 5 x + 6 = 0"
-            toExpr = "Solution[x, {-3, -2}]"
+            toExpr = "SetSolution[x: {-3, -2}]"
             explanation {
                 key = EquationsExplanation.SolveEquationByFactoring
             }
@@ -45,7 +45,7 @@ class EquationsByFactoringTest {
             }
 
             step {
-                toExpr = "Solution[x, {-3, -2}]"
+                toExpr = "SetSolution[x: {-3, -2}]"
                 explanation {
                     key = EquationsExplanation.SolveEquationUnion
                 }
@@ -59,7 +59,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "x + 2 = 0"
-                        toExpr = "Solution[x, {-2}]"
+                        toExpr = "SetSolution[x: {-2}]"
                         explanation {
                             key = EquationsExplanation.SolveLinearEquation
                         }
@@ -75,7 +75,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "x + 3 = 0"
-                        toExpr = "Solution[x, {-3}]"
+                        toExpr = "SetSolution[x: {-3}]"
                         explanation {
                             key = EquationsExplanation.SolveLinearEquation
                         }
@@ -84,7 +84,7 @@ class EquationsByFactoringTest {
 
                 task {
                     taskId = "#3"
-                    startExpr = "Solution[x, {-3, -2}]"
+                    startExpr = "SetSolution[x: {-3, -2}]"
                     explanation {
                         key = EquationsExplanation.CollectSolutions
                     }
@@ -100,7 +100,7 @@ class EquationsByFactoringTest {
 
         check {
             fromExpr = "[x ^ 2] - 4 = 0"
-            toExpr = "Solution[x, {-2, 2}]"
+            toExpr = "SetSolution[x: {-2, 2}]"
             explanation {
                 key = EquationsExplanation.SolveEquationByFactoring
             }
@@ -123,7 +123,7 @@ class EquationsByFactoringTest {
 
             step {
                 fromExpr = "x - 2 = 0 OR x + 2 = 0"
-                toExpr = "Solution[x, {-2, 2}]"
+                toExpr = "SetSolution[x: {-2, 2}]"
                 explanation {
                     key = EquationsExplanation.SolveEquationUnion
                 }
@@ -137,7 +137,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "x - 2 = 0"
-                        toExpr = "Solution[x, {2}]"
+                        toExpr = "SetSolution[x: {2}]"
                         explanation {
                             key = EquationsExplanation.SolveLinearEquation
                         }
@@ -153,7 +153,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "x + 2 = 0"
-                        toExpr = "Solution[x, {-2}]"
+                        toExpr = "SetSolution[x: {-2}]"
                         explanation {
                             key = EquationsExplanation.SolveLinearEquation
                         }
@@ -162,7 +162,7 @@ class EquationsByFactoringTest {
 
                 task {
                     taskId = "#3"
-                    startExpr = "Solution[x, {-2, 2}]"
+                    startExpr = "SetSolution[x: {-2, 2}]"
                     explanation {
                         key = EquationsExplanation.CollectSolutions
                     }
@@ -178,7 +178,7 @@ class EquationsByFactoringTest {
 
         check {
             fromExpr = "[x ^ 6] = 5 [x ^ 5] - 3 [x ^ 4]"
-            toExpr = "Solution[x, {0, [5 - sqrt[13] / 2], [5 + sqrt[13] / 2]}]"
+            toExpr = "SetSolution[x: {0, [5 - sqrt[13] / 2], [5 + sqrt[13] / 2]}]"
             explanation {
                 key = EquationsExplanation.SolveEquationByFactoring
             }
@@ -208,7 +208,7 @@ class EquationsByFactoringTest {
             }
 
             step {
-                toExpr = "Solution[x, {0, [5 - sqrt[13] / 2], [5 + sqrt[13] / 2]}]"
+                toExpr = "SetSolution[x: {0, [5 - sqrt[13] / 2], [5 + sqrt[13] / 2]}]"
                 explanation {
                     key = EquationsExplanation.SolveEquationUnion
                 }
@@ -222,7 +222,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "[x ^ 4] = 0"
-                        toExpr = "Solution[x, {0}]"
+                        toExpr = "SetSolution[x: {0}]"
                         explanation {
                             key = EquationsExplanation.SolveEquationUsingRootsMethod
                         }
@@ -238,7 +238,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "[x ^ 2] - 5 x + 3 = 0"
-                        toExpr = "Solution[x, {[5 - sqrt[13] / 2], [5 + sqrt[13] / 2]}]"
+                        toExpr = "SetSolution[x: {[5 - sqrt[13] / 2], [5 + sqrt[13] / 2]}]"
                         explanation {
                             key = EquationsExplanation.SolveQuadraticEquationUsingQuadraticFormula
                         }
@@ -247,7 +247,7 @@ class EquationsByFactoringTest {
 
                 task {
                     taskId = "#3"
-                    startExpr = "Solution[x, {0, [5 - sqrt[13] / 2], [5 + sqrt[13] / 2]}]"
+                    startExpr = "SetSolution[x: {0, [5 - sqrt[13] / 2], [5 + sqrt[13] / 2]}]"
                     explanation {
                         key = EquationsExplanation.CollectSolutions
                     }
@@ -263,7 +263,7 @@ class EquationsByFactoringTest {
 
         check {
             fromExpr = "[x ^ 6] - [x ^ 2] = 0"
-            toExpr = "Solution[x, {-1, 0, 1}]"
+            toExpr = "SetSolution[x: {-1, 0, 1}]"
             explanation {
                 key = EquationsExplanation.SolveEquationByFactoring
             }
@@ -285,7 +285,7 @@ class EquationsByFactoringTest {
             }
 
             step {
-                toExpr = "Solution[x, {-1, 0, 1}]"
+                toExpr = "SetSolution[x: {-1, 0, 1}]"
                 explanation {
                     key = EquationsExplanation.SolveEquationUnion
                 }
@@ -299,7 +299,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "[x ^ 2] = 0"
-                        toExpr = "Solution[x, {0}]"
+                        toExpr = "SetSolution[x: {0}]"
                         explanation {
                             key = EquationsExplanation.SolveEquationUsingRootsMethod
                         }
@@ -315,7 +315,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "x - 1 = 0"
-                        toExpr = "Solution[x, {1}]"
+                        toExpr = "SetSolution[x: {1}]"
                         explanation {
                             key = EquationsExplanation.SolveLinearEquation
                         }
@@ -331,7 +331,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "x + 1 = 0"
-                        toExpr = "Solution[x, {-1}]"
+                        toExpr = "SetSolution[x: {-1}]"
                         explanation {
                             key = EquationsExplanation.SolveLinearEquation
                         }
@@ -347,7 +347,7 @@ class EquationsByFactoringTest {
 
                     step {
                         fromExpr = "[x ^ 2] + 1 = 0"
-                        toExpr = "Solution[x, {}]"
+                        toExpr = "SetSolution[x: {}]"
                         explanation {
                             key = EquationsExplanation.SolveEquationUsingRootsMethod
                         }
@@ -356,7 +356,7 @@ class EquationsByFactoringTest {
 
                 task {
                     taskId = "#5"
-                    startExpr = "Solution[x, {-1, 0, 1}]"
+                    startExpr = "SetSolution[x: {-1, 0, 1}]"
                     explanation {
                         key = EquationsExplanation.CollectSolutions
                     }